[wip] remove unnecessary code
